Refactor a bunch of stuff
