+++ src/wrapper/controllers/scene-controller.tsx
@@ -1,9 +1,10 @@
 import { NodeController } from "~/wrapper/controllers/node-controller";
 import { ConnectionController } from "./connection-controller";
 import { NodeTypeFile } from "~/wrapper/helpers/node-type-file";
-import { ConnectionSceneData, MinimalNodeSceneData, NodeSceneData, NodeSceneFile, SceneData } from "~/wrapper/helpers/node-scene-file";
-import { GraphNode } from "../nodes/graph-node";
-import { ActionController } from "~/network/controllers/action-controller";
+import { ConnectionSceneData, NodeSceneData, NodeSceneFile, SceneData } from "~/wrapper/helpers/node-scene-file";
+import { ActionController } from "~/network/controllers/actions/action-controller";
+import { NodeSceneRequestData } from "~/network/websocket/request-types";
+import { NodeActionUtils } from "~/network/controllers/actions/node-actions";
 
 export class SceneController {
     node_type_reader: NodeTypeFile;
@@ -30,47 +31,33 @@ export class SceneController {
         this._load_node_types(node_types_path);
         this._load_node_scene(scene_path);
     }
-
-    public gen_scene_data(): SceneData {
-        const node_types_data = this.node_type_reader.raw_data;
-        const node_types_id = this.node_type_reader.node_types_id;
-        const node_types_version = this.node_type_reader.node_types_version;
-
-        let scene_nodes: Map<string, NodeSceneData> = new Map();
-        this.node_controller.nodes.forEach((node, idx) => {
-            scene_nodes.set(`node_${node.id}`, {
-                type: node.type_name,
-                position: node.pos,
-                size: node.rect.size,
-                data: node.node_data.map_parameters()
-            });
-        });
-
-        let scene_connections: Map<string, ConnectionSceneData> = new Map();
-        this.connection_controller.connections.forEach((conn, id) => {
-            const input_path: string = NodeSceneFile.make_slot_path(conn.input_slot);
-            const output_path: string = NodeSceneFile.make_slot_path(conn.output_slot);
-            scene_connections.set(`connection_${conn.uid}`, {
-                from: output_path,
-                to: input_path
-            });
-        });
-
-        const scene_data: SceneData = {
-            node_types_id: node_types_id != null ? node_types_id : "unknown",
-            node_types_version: node_types_version,
-            nodes: scene_nodes,
-            connections: scene_connections
-        }
-
-        return scene_data;
-    }
-
+    
     public save_scene(): SceneData {
-        const scene_data = this.gen_scene_data()
+        const scene_data = SceneUtils.gen_scene_data(this)
         this.node_scene_reader.save_data_to_file(scene_data);
         return scene_data;
     }
+
+    public _clear_scene() {
+        this.node_controller.clear();
+        this.connection_controller.clear();
+    }
+
+        public load_node_type_data(type_data: any) {
+        this.node_type_reader.load_type_data(type_data);
+        this.node_controller.load_node_types(this.node_type_reader);
+    }
+
+    public load_scene_data(scene_data: any) {
+        if (scene_data && Object.entries(scene_data).length > 0) {
+            this.node_scene_reader.load_from_json_data(scene_data);
+            this.node_scene_reader.swap_virtual_data();
+        }
+        
+        this._clear_scene();
+        this._parse_loaded_node_scene();
+    }
+
 
     protected _load_node_scene(file_path: string) {
         this.node_scene_reader._load_file_path_async(file_path).then(
@@ -105,19 +92,21 @@ export class SceneController {
         return true;
     }
 
-    public _clear_scene() {
-        this.node_controller.clear();
-        this.connection_controller.clear();
+    protected _load_node_types(file_path: string) {
+        this.node_type_reader._load_file_async(file_path).then(
+            () => {
+                this.node_controller.load_node_types(this.node_type_reader);
+            }
+        );
     }
 
-    // 
     protected _parse_loaded_node_scene() {
         console.log(this.node_scene_reader.scene_data)
         if (this.node_scene_reader.scene_data == null) {
             return;
         }
 
-        let nodes: {[uid: string]: MinimalNodeSceneData;} = {};
+        let nodes: NodeSceneRequestData = {};
         this.node_scene_reader.scene_data.nodes.forEach((node_data: NodeSceneData, node_key: string) => {
             const constructor = this.node_type_reader.get_constructor(node_data.type);
             if (constructor) {
@@ -128,7 +117,7 @@ export class SceneController {
                 }
             }
         });
-        this.action_controller.request_add_nodes(nodes);
+        NodeActionUtils.request_add_nodes(nodes, this.action_controller);
 
         this.node_scene_reader.scene_data.connections.forEach((conn_data: ConnectionSceneData, conn_uid: string) => {
             const node_a_path = NodeSceneFile.parse_node_path(conn_data.from);
@@ -158,37 +147,41 @@ export class SceneController {
             );
         });
     }
+}
 
-    protected _load_node_types(file_path: string) {
-        this.node_type_reader._load_file_async(file_path).then(
-            () => {
-                this.node_controller.load_node_types(this.node_type_reader);
-            }
-        );
-    }
-    
-    public load_node_type_data(type_data: any) {
-        this.node_type_reader.load_type_data(type_data);
-        this.node_controller.load_node_types(this.node_type_reader);
-    }
+export class SceneUtils {
+    public static gen_scene_data(scene_controller: SceneController): SceneData {
+        const node_types_data = scene_controller.node_type_reader.raw_data;
+        const node_types_id = scene_controller.node_type_reader.node_types_id;
+        const node_types_version = scene_controller.node_type_reader.node_types_version;
 
-    public load_scene_data(scene_data: any) {
-        // this.node_scene_reader.clear_data();
-        if (scene_data && Object.entries(scene_data).length > 0) {
-            this.node_scene_reader.load_from_json_data(scene_data);
-            this.node_scene_reader.swap_virtual_data();
-        }
-        
-        this._clear_scene();
-        this._parse_loaded_node_scene();
-    }
-
-    public add_nodes_unsynced(nodes_data: {[uid: string]: MinimalNodeSceneData}) {
-        Object.entries(nodes_data).forEach(([uid, node_data]) => {
-            this.node_controller.add_new_node(
-                "", node_data.position, node_data.type,
-                uid 
-            );
+        let scene_nodes: Map<string, NodeSceneData> = new Map();
+        scene_controller.node_controller.nodes.forEach((node, idx) => {
+            scene_nodes.set(`node_${node.id}`, {
+                type: node.type_name,
+                position: node.pos,
+                size: node.rect.size,
+                data: node.node_data.map_parameters()
+            });
         });
+
+        let scene_connections: Map<string, ConnectionSceneData> = new Map();
+        scene_controller.connection_controller.connections.forEach((conn, id) => {
+            const input_path: string = NodeSceneFile.make_slot_path(conn.input_slot);
+            const output_path: string = NodeSceneFile.make_slot_path(conn.output_slot);
+            scene_connections.set(`connection_${conn.uid}`, {
+                from: output_path,
+                to: input_path
+            });
+        });
+
+        const scene_data: SceneData = {
+            node_types_id: node_types_id != null ? node_types_id : "unknown",
+            node_types_version: node_types_version,
+            nodes: scene_nodes,
+            connections: scene_connections
+        }
+
+        return scene_data;
     }
 }
